requirements.txt
Flask==2.1.2
numpy==1.21.2
Pillow==8.3.2
tensorflow==2.6.0
